Check authentication before delete a blog

+++ whisky-ng-server/api/v1/index.js
@@ -76,6 +76,9 @@ router.post('/blog-posts', (req, res) => {
 });
 
 router.delete('/blog-posts/:id', (req, res) => {
+	if (!req.isAuthenticated()) {
+		return res.status(401).json({ result: 'Ko', msg: 'Not authorized to delete a blog'});
+	}
 	const id = req.params.id;
 	Blogpost.findByIdAndDelete(id, (err, blogPost) => {
 		if(err)
@@ -91,6 +94,10 @@ router.delete('/blog-posts/:id', (req, res) => {
 });
 
 router.delete('/blog-posts', (req, res) => {
+	if (!req.isAuthenticated()) {
+		return res.status(401).json({ result: 'Ko', msg: 'Not authorized to delete a blog'});
+	}
+
 	// url de la forme: http://localhost:3000/api/v1/blog-post?ids=1z3es,qa12ess	 
 	const ids = req.query.ids;
 	console.log('ids', ids);
